refactor(movement): 인터페이스에서 클래스로 변경
