--- src/main/java/janggi/domain/movement/StraightMovement.java
@@ -1,64 +1,0 @@
-package janggi.domain.movement;
-
-import janggi.domain.Position;
-import janggi.domain.board.BoardMediator;
-import janggi.domain.piece.Piece;
-import java.util.ArrayList;
-import java.util.List;
-
-public class StraightMovement implements Movement {
-    private final BoardMediator boardMediator;
-    private final int maxDistance;
-    private final Direction direction;
-
-    public StraightMovement(int maxDistance, Direction direction, BoardMediator boardMediator) {
-        this.maxDistance = maxDistance;
-        this.direction = direction;
-        this.boardMediator = boardMediator;
-    }
-
-    @Override
-    public boolean canReach(final Position from) {
-        for (int distance = 1; distance <= maxDistance; distance++) {
-            Position to = from.calculateNext(distance, direction);
-            if (boardMediator.existsInPosition(to)) {
-                return false;
-            }
-        }
-        return true;
-    }
-
-    @Override
-    public Position calculateDestination(final Position from) {
-        Piece fromPiece = boardMediator.getPieceInPosition(from);
-        for (int distance = 1; distance <= maxDistance; distance++) {
-            Position to = from.calculateNext(distance, direction);
-            if (boardMediator.existsInPosition(to)) {
-                Piece toPiece = boardMediator.getPieceInPosition(to);
-                if (toPiece.belongsToTeam(fromPiece.getTeamType())) {
-                    return from.calculateNext(distance - 1, direction);
-                }
-                return from.calculateNext(distance, direction);
-            }
-        }
-        return from.calculateNext(maxDistance, direction);
-    }
-
-    @Override
-    public List<Position> calculateTraces(final Position from) {
-        final Piece fromPiece = boardMediator.getPieceInPosition(from);
-        final List<Position> traces = new ArrayList<>();
-        for (int distance = 1; distance <= maxDistance; distance++) {
-            Position to = from.calculateNext(distance, direction);
-            if (boardMediator.existsInPosition(to)) {
-                Piece toPiece = boardMediator.getPieceInPosition(to);
-                if (!toPiece.belongsToTeam(fromPiece.getTeamType())) {
-                    traces.add(to);
-                }
-                return traces;
-            }
-            traces.add(to);
-        }
-        return traces;
-    }
-}
